hotfix

+++ src/system.rs
@@ -204,7 +204,7 @@ pub fn check_turbo_enabled() -> Result<bool, Error> {
 
 pub fn check_bat_cond() -> Result<f32, Error> {
     let bat_cond_calc: f32 =
-        (read_bat_energy_full(false).unwrap() / read_bat_energy_full(true).unwrap()) as f32;
+        read_bat_energy_full(false).unwrap() as f32 / read_bat_energy_full(true).unwrap() as f32;
     Ok(bat_cond_calc)
 }
